# Lab2/openapi.yaml
openapi: 3.0.3
info:
  title: Computer Store
  version: 1.0.0
...
paths:
  /products:
    get:
      tags:
        - products
      summary: Get all productss
      responses:
        '200':
          description: A list of productss
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/products'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/products'
        '500':
          description: Status internal server error
    post:
      tags:
        - products
      summary: Add a new products to the store
      description: Add a new products to the store
      operationId: addproducts
      requestBody:
        description: Create a new products in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/products'
          application/xml:
            schema:
              $ref: '#/components/schemas/products'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /products/{id}:
    get:
      tags:
        - products
      summary: Get products by id
      operationId: getproductsById
      parameters:
        - name: id
          in: path
          description: ID products
          required: true
          schema:
            type: integer
            format: int64
            description: the id of the products
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/products'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/products'
        '404':
          description: products not found
        '500':
          description: Status internal server error
    put:
      tags:
        - products
      summary: Updates a products in the store with form data
      description: ''
      operationId: updateproducts
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/products'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    patch:
      tags:
        - products
      summary: Partially update a specific products by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Fields to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/products'
      responses:
        '200':
          description: products partially updated successfully
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: products not found
    delete:
      tags:
        - products
      summary: Deletes a products
      description: delete a products
      operationId: deleteproducts
      parameters:
        - name: id
          in: path
          description: products id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /products?name=string:
    get:
      tags:
        - products
      summary: Get a list of productss
      parameters:
        - name: name
          in: query
          description: Name of products
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of productss
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/products'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/products'
        '500':
          description: Status internal server error
...
  /orders:
    get:
      tags:
        - orders
      summary: Get all orderss
      description: Get users all orderss
      operationId: GetAllorderss
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/orders'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/orders'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    post:
      tags:
        - orders
      summary: Add a new orders
      description: Add a new orders
      operationId: addorders
      requestBody:
        description: Create a new orders
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orders'
          application/xml:
            schema:
              $ref: '#/components/schemas/orders'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /orders/{id}:
    get:
      tags:
        - orders
      summary: Get orders by id
      description: Get orders by id
      operationId: GetordersById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/orders'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/orders'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - orders
      summary: Updates status orders
      description: ''
      operationId: updateorders
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orders'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - orders
      summary: Deletes a orders
      description: delete a orders
      operationId: deleteorders
      parameters:
        - name: id
          in: path
          description: users id to orders
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /orders?users_id=int:
    get:
      tags:
        - orders
      summary: Get orders by users_id
      description: Get orders by users_id
      operationId: GetordersByIdusers
      parameters:
        - name: users_id
          in: query
          description: The id users that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/orders'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/orders'
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
...
components:
  schemas:
    products:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Dell
        price:
          type: integer
          example: 100
        quantity:
          type: integer
          example: 10
        description:
          type: string
          example: In 2020
      xml:
        name: products
...
    orders:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        users_id:
          type: integer
          format: int64
          example: 1
        status:
          type: string
          enum:
            - pending
            - delivering
            - delivered
          example: Delivered
        data_created:
          type: string
          format: date-time
          example: '2024-09-20'
      xml:
        name: orders
...